Fixes a Windows specific bottle-duping bug.
Compiler Optimizations on Windows were causing GettingItemInBottle
to always return true, rather than checking the gettingBottledItem
variable each time it was called. As a result, if you catch a Poe,
it would give you a new bottle instead of putting a Poe in your
existing bottle. The fix for now was simply turning optimization
off for that function.