add -DNO_PULSE if NO_PULSE=ON when building soh
OTRGlobals.cpp includes AudioPlayer.h from libultraship, which pulls in
PulseAudio headers. This points to the build options probably belonging in the
root CMakeLists rather than in libultraship's. It works here, but it should
definitely be moved to a more common location.

This works around building soh via a cross-compilation toolchain and/or on a
system that does not have PulseAudio.