Track time rate in any area
Now the rate of time slider will be able to display the current rate of time in any area.
Various Time Cheats
The slider will tell you the current time of day.
Added Skybox Updates
Now, when a person selects to change the time, the skybox will update.
Added a clock to cheats
Added a clock to tell what the time currently is.
Logic to display time rate
Added logic to show the current time rate on the time rate modifier slider.