Added more spots for quick ocarina
I found that MANY spots are activated with this area of code (the triforces on the ground in temples, entrance to Darunia's room, Fairy Fountains, etc.) Exceptions not covered yet (that I know of) include Darunia himself, the Song of Storms grave, and possibly Time blocks. I will test more and see what else to add.