Fixed issues
Bridge now will properly lower during opening cut scene.
Since exiting the marketplace at night is only able to be done if it is rendered as daytime, that pre-rendered scene is the daytime variant. However, to accommodate this, I've disabled back ground music so it's less jarring of a transition.