Converted checking functionality to send UpdateCheckData packet instead of trying to use the receiver's check update system to capture collected items.
Streamlined the location checking to utilize a vector to store check areas that the game just always checks until it finds a newly collected one. This will still present an issue when using the new `Teleport To` functionality in Anchor v8 if the player teleports to another already in a grotto and then collects a check, if grotto shuffle is not on.

Added itemsReceived to prevent initiation of checking checks on remote item reception.