Reinstate "gAltLinkEquip" cvar, now as part of the CanUse functions.
Adds cvar checks for "gAltLinkEquip" to the `Player_CanUseNewLoadingMethod` family of functions that were once intended to replace it.

+ (z_player_lib.c) Partial formatting pass.
+ Repurpose the ghost instances of checking against gAltLinkEquip in z_arms_hook.c that Luna didn't get around to removing before. (This may need further attention.)

Explanation:

The "gAltLinkEquip" cvar was once in this repo but replaced with the `Player_CanUseNewLoadingMethod` family of functions. (see commit `51f4fa24`)
The original author's intent was to make this seamless to the player without altering vanilla behavior. For a variety of reasons this doesn't really bear out, and as I understand it, a cvar thusly still needs to be present to be able to override it.

Unlike before, this cvar now defaults to true and is otherwise an invisible option to the general player.
If turned off in the developer console, the game should fall back on vanilla behavior. At least for these aspects.

This hopefully goes some way to addressing the issue of Alternate Equipment Loading not being clearly separate from vanilla behavior, while also keeping the intent to make using mods that require this functionality seamless for players.