Reduce Minimum Internal Resolution Scalar to 0.1f.
This setting dates back to https://github.com/HarbourMasters/Shipwright/pull/3130 , which also adds an Advanced Resolution Editor (that doesn't appear to be available on the Switch).   This value appears to be arbitrarily set.

Reducing this minimum value appears to be the easiest way to enable N64-like resolutions with widescreen enabled on the Switch platform.   When forced via the config.json, it snaps back to 0.5f whenever that menu is navigated to.

I've noticed this complaint in some Reddit comments as well.   This should also help on non-switch, High-DPI screens as well.