Separate our custom save context data.
Separates the save context data added for ship from z64save.h using Object Extensions. This came with several include headers that had to be added, apparently cvar_prefixes.h was being included in one of the files z64save.h was including before that change that got them into most of the files it was needed in previously, a significant chunk of these changes is just including that file in more places.